+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Enzyme: Upgrade to the latest `v3.10.0` version and pull in Flow library changes (#543)
 - Eslint: Bump all related packages/plugins to current latest version (#544)
+- Button: add new `textColor` prop to allow overriding of text color for buttons (#545)
 
 ### Patch